requirements.txt
pytest~=7.1.3
allure-pytest
python-dotenv==0.20.0
Appium-Python-Client~=2.6.1
selenium~=4.2.0
requests
pydantic~=1.10.2
black>=22.8.0
flake8>=5.0.4
isort>=5.10.1
pre-commit>=2.20.0
vulture>=2.5
pre-commit-hooks>=4.3.0
Faker>=14.2.0
enum
